+++ ts/features/itwallet/common/saga/index.ts
@@ -1,7 +1,10 @@
 import { SagaIterator } from "redux-saga";
-import { fork, select } from "typed-redux-saga/macro";
+import { fork, put, select } from "typed-redux-saga/macro";
 import { isItWalletTestEnabledSelector } from "../../../../store/reducers/persistedPreferences";
+import { trialSystemActivationStatus } from "../../../trialSystem/store/actions";
 import { watchItwIdentificationSaga } from "../../identification/saga";
+import { handleWalletCredentialsRehydration } from "../../credentials/saga/handleWalletCredentialsRehydration";
+import { ITW_TRIAL_ID } from "../utils/itwTrialUtils";
 
 export function* watchItwSaga(): SagaIterator {
   const isItWalletTestEnabled: ReturnType<
@@ -13,5 +16,9 @@ export function* watchItwSaga(): SagaIterator {
     return;
   }
 
+  yield* fork(handleWalletCredentialsRehydration);
   yield* fork(watchItwIdentificationSaga);
+
+  // IT Wallet trial status refresh
+  yield* put(trialSystemActivationStatus.request(ITW_TRIAL_ID));
 }